Create a new quiz To create a new quiz, the client needs to send a JSON as the request's body via POST to /api/quizzes. The JSON should contain the four fields: title (a string), text (a string), options (an array of strings) and answer (integer index of the correct option). At this moment, all the keys are optional.
Here is a new JSON quiz as an example:

{
  "title": "The Java Logo",
  "text": "What is depicted on the Java logo?",
  "options": ["Robot","Tea leaf","Cup of coffee","Bug"],
  "answer": 2
}
The answer equals 2 corresponds to the third item from the options array ("Cup of coffee").

The server response is a JSON with four fields: id, title, text and options. Here is an example.

{
  "id": 1,
  "title": "The Java Logo",
  "text": "What is depicted on the Java logo?",
  "options": ["Robot","Tea leaf","Cup of coffee","Bug"]
}
The id field is a generated unique integer identifier for the quiz. Also, the response may or may not include the answer field depending on your wishes. This is not very important for this operation.

At this moment, it is admissible if a creation request does not contain some quiz data. In the next stages, we will improve the service to avoid some server errors.

Get a quiz by id
To get a quiz by id, the client sends the GET request to /api/quizzes/{id}.

Here is a response example:

{
  "id": 1,
  "title": "The Java Logo",
  "text": "What is depicted on the Java logo?",
  "options": ["Robot","Tea leaf","Cup of coffee","Bug"]
}
The response must not include the answer field, otherwise, any user will be able to find the correct answer for any quiz.
If the specified quiz does not exist, the server should return the 404 (Not found) status code.

Get all quizzes
To get all existing quizzes in the service, the client sends the GET request to /api/quizzes.

The response contains a JSON array of quizzes like the following:

[
  {
    "id": 1,
    "title": "The Java Logo",
    "text": "What is depicted on the Java logo?",
    "options": ["Robot","Tea leaf","Cup of coffee","Bug"]
  },
  {
    "id": 2,
    "title": "The Ultimate Question",
    "text": "What is the answer to the Ultimate Question of Life, the Universe and Everything?",
    "options": ["Everything goes right","42","2+2=4","11011100"]
  }
]
The response must not include the answer field, otherwise, any user will be able to find the correct answer for any quiz.
If there are no quizzes, the service returns an empty JSON array: [].

In both cases, the status code is 200 (OK).

Solving a quiz
To solve the quiz, the client sends a POST request to /api/quizzes/{id}/solve and passes the answer parameter in the content. This parameter is the index of a chosen option from options array. As before, it starts from zero.

The service returns a JSON with two fields: success (true or false) and feedback (just a string). There are three possible responses.

If the passed answer is correct (e.g., POST to /api/quizzes/1/solve with content answer=2):
{"success":true,"feedback":"Congratulations, you're right!"}
If the answer is incorrect (e.g., POST to /api/quizzes/1/solve with content answer=1):
{"success":false,"feedback":"Wrong answer! Please, try again."}
If the specified quiz does not exist, the server returns the 404 (Not found) status code.
You can write any other strings in the feedback field, but the names of fields and the true/false values must match this example.